Refactor of files.
content_layout -> layout and layout_width.
