+++ widgets/table/layout.go
@@ -1,0 +1,40 @@
+// Copyright 2019 Google Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package table
+
+// layout.go stores layout calculated for a canvas size.
+
+import (
+	"errors"
+	"image"
+)
+
+// contentLayout determines how the content gets placed onto the canvas.
+type contentLayout struct {
+	// lastCvsAr is the are of the last canvas the content was drawn on.
+	// This is image.ZR if the content hasn't been drawn yet.
+	lastCvsAr image.Rectangle
+
+	// columnWidths are the widths of individual columns in the table.
+	columnWidths []columnWidth
+
+	// Details about HV lines that are the borders.
+}
+
+// newContentLayout calculates new layout for the content when drawn on a
+// canvas represented with the provided area.
+func newContentLayout(content *Content, cvsAr image.Rectangle) (*contentLayout, error) {
+	return nil, errors.New("unimplemented")
+}
